requirements.txt
selenium==4.4.3
webdriver-manager==3.8.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1